[45] Make navbar smaller on scroll
