+++ src/styles/_variables.scss
@@ -12,4 +12,5 @@ $bp-min-desktop: 769px;
 $max-page-width: 1024px;
 $page-inner-padding: 1rem;
 $header-height: 5vh;
+$header-height-small: 3.5vh;
 $footer-height: 5vh;
